Move testhelpers.ts to seperate folder
